requirements/production.txt
# Everything needed in production

wheel

# Flask
Flask
Flask-RESTful
blinker

# Database
Flask-SQLAlchemy
psycopg2
GeoAlchemy2

# Migrations
Flask-Migrate

# Deployment
gunicorn
gevent
